Install OpenCV via pip to avoid protobuf issues
The old version won't work on Ubuntu 16.04 which has following problems when running `learner.py`:

This program was compiled against version 3.5.1 of the Protocol Buffer runtime library, which is not compatible with the installed version (3.8.0).  Contact the program author for an update.  If you compiled the program yourself, make sure that your headers are from the same version of Protocol Buffers as your link-time library.